Rename payment to transaction, and payment_view to payment
Reasoning:
- Transaction will provide aptly named transaction services between any
  two parties
- Payment is a RESTful app that provides actual payment capabilities
  with the outside world
- Decoupled Transaction from Core app
