+++ backend/app/urls.py
@@ -31,5 +31,5 @@ urlpatterns = [
     ),
     path("user/", include("user.urls")),
     path("", include("loan_profile.urls")),
-    path("payment/", include("payment_view.urls")),
+    path("payment/", include("payment.urls")),
 ]
