+++ backend/payment/admin.py
@@ -1,6 +1,3 @@
 from django.contrib import admin
 
-from payment import models
-
-
-admin.site.register(models.Transaction)
+# Register your models here.
