+++ backend/payment/apps.py
@@ -2,5 +2,5 @@ from django.apps import AppConfig
 
 
 class PaymentConfig(AppConfig):
-    default_auto_field = "django.db.models.BigAutoField"
-    name = "payment"
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'payment'
