--- backend/payment_view/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class PaymentViewConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'payment_view'
